refactor: pkg/common
