+++ pkg/webhook/config.go
@@ -24,7 +24,7 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/bank-vaults/vault-secrets-webhook/common"
+	"github.com/bank-vaults/vault-secrets-webhook/pkg/common"
 )
 
 // VaultConfig represents vault options
